# specs/001-team-health-dashboard/contracts/api.yaml
openapi: 3.0.3
info:
  title: チーム健全性ダッシュボード API
  description: |
    GitHubリポジトリのIssue/PRを分析し、チームの健全性を評価するAPI。

    ## 同期方式
    - 初回: 全件取得
    - 2回目以降: lastSyncedAt以降の更新分のみ取得（差分同期）

    ## 表示フィルタ
    - スプリント期間を指定して、該当期間のデータのみ表示可能
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API Routes
paths:
...
  /github/sync:
    post:
      summary: GitHubからデータを同期（差分同期）
      description: |
        初回は全件取得、2回目以降はlastSyncedAt以降の更新分のみ取得。
        fullSync=trueで強制的に全件再取得。
      operationId: syncGitHub
      tags:
        - GitHub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - repositoryId
              properties:
                repositoryId:
                  type: string
                fullSync:
                  type: boolean
                  default: false
                  description: trueの場合、lastSyncedAtを無視して全件再取得
      responses:
        '200':
          description: 同期完了
          content:
            application/json:
              schema:
                type: object
                properties:
                  issuesSynced:
                    type: integer
                    description: 同期されたIssue数
                  prsSynced:
                    type: integer
                    description: 同期されたPR数
                  isFullSync:
                    type: boolean
                    description: 全件同期かどうか
                  lastSyncedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: GitHub APIレート制限
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  retryAfter:
                    type: string
                    format: date-time
                    description: 次回リクエスト可能時刻
...
components:
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: 認証が必要
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Firebase ID Token
security:
  - BearerAuth: []
...